Move PVC logic from tests/utils to libstorage/pvc
