+++ tests/template_test.go
@@ -268,7 +268,7 @@ var _ = Describe("[Serial][sig-compute]Templates", func() {
 		Context("[rfe_id:273][crit:medium][vendor:cnv-qe@redhat.com][level:component]with RHEL Template", func() {
 			BeforeEach(func() {
 				checks.SkipIfNoRhelImage(virtClient)
-				tests.CreatePVC(tests.OSRhel, "15Gi", libstorage.Config.StorageClassRhel, true)
+				libstorage.CreatePVC(tests.OSRhel, "15Gi", libstorage.Config.StorageClassRhel, true)
 				AssertTemplateSetupSuccess(vmsgen.GetTestTemplateRHEL7(), nil)()
 			})
